notify clientconn about resolver errors via  clientConn.ReportError()
Since grpc-go 1.25.0 resolvers can notify the ClientConn via ReportError() that
an error happened.
This allows failFast clientConns to fail on resolver errors.
It also simplifies the error handling that is required in the resolver.
When the clientConn is notified about an error, the loadbalancer will call the
ResolveNow() function periodically with an exponential backoff.

Call clientConn.ReportError() on resolver errors.
This allows to remove the logic to retry on resolver errors with a backoff
timeout.