support resolving to unhealthy services if no healthy ones are available
The dial URL supports a new parameter 'health' to allow filtering services by
their health status.

The parameter support the values 'healthy' and 'fallbackToUnhealthy'.
The default is 'healthy' which is the same behavior then before, only resolve to
services that have health-status "passing".

If the health parameter is set to 'fallbackToUnhealthy', the resolver resolves
to services with any health status if none exist with a passing health-check.