make conditions for sleeping between queries more strict
Previously the resolver delayed the query, if consul responded with the same
data but a different waitindex.
This can e.g. happen if we only query for healthy services, an unhealthy service
was registered and then consul responded with a new waitindex but the same data.

Only sleep between queries if the returned waitindex is the same then in the
last query and consul responded in less then 50msec.