fix: resolver delayed next query on fast empty responses unnecessarily
When consul returned no addresses 2x times in a row but different waitIndices
the next query was delayed for 50ms.
The following query should only be delayed if the same addresses and same
WaitIndex is returned then in the previous call.

The cause was that lastWaitIndex was set to the new waitIndex after the query
was returned, before the check.