fix: http2 authority header set to bogus value
The resolver did not set the ServerName field in the resolver.Address structs.
The ServerName field sets, according to the docs, the http2 authority header
field.
When Address.ServerName is empty the grpc-go library parses the authority part
of the url provided to grpc.Dial() and uses it as authority.
Because we a custom URL scheme
(consul://consul-server/serviceName[?<OPT>[&<OPT>]...]), the grpc-go lib parsed
wrongly as authority the part 'serviceName[?<OPT>[&<OPT>]...]` used it in the
http2 header.

Go GRPC-Servere were ignoring this bogus authority field but other
implementations like node's grpc/grpc-js were failing grpc-requests with a
PROTOCOL_ERROR.

Set Address.ServerName to the same then address.