retry after pause when UpdateState returns an error
Since grpc-go 1.38 the UpdateState() call of the ClientConn can return an error.

When UpdateState() returns an error the grpcconsulresolver now polls consul and
retries calling UpdateState() again after a backoff pause.
This is analogue to how the dns resolver that is part of the grpc-go package
handles UpdateState() errors.

Instead of using a timer to pause before polling consul again, the WaitTime for
the consul query is set to the amount of the backoff timeout.
This has the advantage that if the address is consul is updated, the wait is
interrupted and UpdateState() is called again with the new addresses.

The backoff implementation was restored from the commit
87eb606b2cd7c3d73a70008b3953742fb6768e05.