ignore errors returned by UpdateState
When UpdateState returns an error, the same address was reported again to
UpdateState() after a backoff timeout expired.
If the address to that consul resolved the service changed in the meantime the
new one was passed to UpdateState().
Calling UpdateState() with the same address, when it returned an error, is not
needed.

According to the explanation in https://github.com/grpc/grpc-go/issues/5048
there is no scenario where UpdateState() would fail because of a temporary
condition and succeed when called with the same address later.
The returned error by UpdateState() can be ignored for watch based resolver.

This reverts 99% of the changes of the commit
f1f1a15 ("retry after pause when UpdateState returns an error")
Compared to the previous
commit 1188d19 ("update google.golang.org/grpc to version 1.38.0") only the
method signature of UpdateState() is changed to return an error and the resolver
now logs a info message with verbosity lvl 2 if UpdateState returns an error.