Update grpc-go to 1.53 & fix that resolver opts were not passed
The minimum required grpc-go version is changed to 1.53.0.
1.53.0 contains a breaking resolver.Target interface change. The Target.Endpoint
field was replaced a Target.Endpoint method().
The Target.Scheme and Target.Authority were deprecated since v1.42.0, instead
Target.URLI should be used instead. The grpc-consul-resolver is adapted to those
changes.

When using grpc-go >=v1.42.0 with grpc-consul-resolver passing options via
query-strings through grpc.Dial() was broken. The resolver was expecting that
target.Endpoint contains the query-string but since grpc-go PR #4817[^1] it was
not the case anymore.

[^1]: https://github.com/grpc/grpc-go/pull/4817