Utilize url.URL when parsing the target url
Since grpc-go 1.42.0  resolver.Target contains an url.URL struct.
Use the utility methods of the URL struct to retrieve the url path and parse the
query-parameter, instead of implement it on our own.

The behavior when an option is defined multiple in a query string is now also
documented in the package doc and tested in a unit-test.