prevent calling ReportError with the same error in a row
The resolver already prevents that UpdateState is called for the same
address multiples times in a row, do the same when reporting errors.
Only call ReportError() if the same error hasn't been reported in the
last call and UpdateState hasn't been called last.

This prevents that we unnecessary report the same state multiple times
and trigger the same events for the connection.

The code is refactored to store the last reported state in the
consulResolver struct and reporting addresses and errors is moved to
separate methods.